Creation of basis functions

fourier_2d() and polynomial_2d() creates a list of basis functions to be used as the fixed effect in fit_ghm.

fourier_2d(max_freqs, lattice_dim)

polynomial_2d(poly_deg, lattice_dim)

Arguments

max_freqs
A length 2 numeric vector with maximum frequencies considered (x-axis and y-axis direction, respectively).

lattice_dim
A length 2 numeric vector with lattice dimensions (N,M) to be used.

poly_deg
A length 2 numeric vector with degrees of the bivariate polynomial to be considered.

Value

A list of functions.

Details

fourier_2d() is for 2-dimensional Fourier transform.

Examples

fourier_2d(c(10,10), dim(Z_potts))

#> $`2`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a741760>
#> 
#> $`3`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a741568>
#> 
#> $`4`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a741370>
#> 
#> $`5`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a741178>
#> 
#> $`6`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a740f80>
#> 
#> $`7`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91ba28>
#> 
#> $`8`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91b830>
#> 
#> $`9`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91b638>
#> 
#> $`10`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91b440>
#> 
#> $`11`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91b248>
#> 
#> $`12`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91b050>
#> 
#> $`13`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91ae58>
#> 
#> $`14`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91ac60>
#> 
#> $`15`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91aa68>
#> 
#> $`16`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91a870>
#> 
#> $`17`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91a678>
#> 
#> $`18`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91a480>
#> 
#> $`19`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91a288>
#> 
#> $`20`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b91a090>
#> 
#> $`21`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b919e98>
#> 
#> $`22`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b919ca0>
#> 
#> $`23`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdeb7e8>
#> 
#> $`24`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdeb5f0>
#> 
#> $`25`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdeb3f8>
#> 
#> $`26`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdeb200>
#> 
#> $`27`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdeb008>
#> 
#> $`28`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdeae10>
#> 
#> $`29`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdeac18>
#> 
#> $`30`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdeaa20>
#> 
#> $`31`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdea828>
#> 
#> $`32`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdea630>
#> 
#> $`33`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdea438>
#> 
#> $`34`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdea240>
#> 
#> $`35`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdea048>
#> 
#> $`36`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bde9e50>
#> 
#> $`37`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bde9c58>
#> 
#> $`38`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfdfa0>
#> 
#> $`39`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfdda8>
#> 
#> $`40`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfdbb0>
#> 
#> $`41`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfd9b8>
#> 
#> $`42`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfd7c0>
#> 
#> $`43`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfd5c8>
#> 
#> $`44`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfd3d0>
#> 
#> $`45`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfd1d8>
#> 
#> $`46`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfcfe0>
#> 
#> $`47`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfcde8>
#> 
#> $`48`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfcbf0>
#> 
#> $`49`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfc9f8>
#> 
#> $`50`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfc800>
#> 
#> $`51`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfc608>
#> 
#> $`52`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfc410>
#> 
#> $`53`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bdfc218>
#> 
#> $`54`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c9720>
#> 
#> $`55`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c9528>
#> 
#> $`56`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c9330>
#> 
#> $`57`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c9138>
#> 
#> $`58`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c8f40>
#> 
#> $`59`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c8d48>
#> 
#> $`60`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c8b50>
#> 
#> $`61`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c8958>
#> 
#> $`62`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c8760>
#> 
#> $`63`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c8568>
#> 
#> $`64`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c8370>
#> 
#> $`65`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c8178>
#> 
#> $`66`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c7f80>
#> 
#> $`67`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c7d88>
#> 
#> $`68`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c7b90>
#> 
#> $`69`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5c7998>
#> 
#> $`70`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5ccc90>
#> 
#> $`71`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cca98>
#> 
#> $`72`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cc8a0>
#> 
#> $`73`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cc6a8>
#> 
#> $`74`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cc4b0>
#> 
#> $`75`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cc2b8>
#> 
#> $`76`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cc0c0>
#> 
#> $`77`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cbec8>
#> 
#> $`78`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cbcd0>
#> 
#> $`79`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cbad8>
#> 
#> $`80`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cb8e0>
#> 
#> $`81`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cb6e8>
#> 
#> $`82`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cb4f0>
#> 
#> $`83`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cb2f8>
#> 
#> $`84`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5cb100>
#> 
#> $`85`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5caf08>
#> 
#> $`86`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39f2f0>
#> 
#> $`87`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39f0f8>
#> 
#> $`88`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39ef00>
#> 
#> $`89`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39ed08>
#> 
#> $`90`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39eb10>
#> 
#> $`91`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39e918>
#> 
#> $`92`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39e720>
#> 
#> $`93`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39e528>
#> 
#> $`94`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39e330>
#> 
#> $`95`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39e138>
#> 
#> $`96`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39df40>
#> 
#> $`97`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39dd48>
#> 
#> $`98`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39db50>
#> 
#> $`99`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39d958>
#> 
#> $`100`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39d760>
#> 
#> $`101`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a39d568>
#> 
#> $`102`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b383110>
#> 
#> $`103`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b382f18>
#> 
#> $`104`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b382d20>
#> 
#> $`105`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b382b28>
#> 
#> $`106`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b382930>
#> 
#> $`107`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b382738>
#> 
#> $`108`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b382540>
#> 
#> $`109`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b382348>
#> 
#> $`110`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b382150>
#> 
#> $`111`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b381f58>
#> 
#> $`112`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b381d60>
#> 
#> $`113`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b381b68>
#> 
#> $`114`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b381970>
#> 
#> $`115`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b381778>
#> 
#> $`116`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b381580>
#> 
#> $`117`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d8e68>
#> 
#> $`118`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d8c70>
#> 
#> $`119`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d8a78>
#> 
#> $`120`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d8880>
#> 
#> $`121`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d8688>
#> 
#> $`2`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd35818>
#> 
#> $`3`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd35620>
#> 
#> $`4`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd35428>
#> 
#> $`5`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd35230>
#> 
#> $`6`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd35038>
#> 
#> $`7`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd34e40>
#> 
#> $`8`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd34c48>
#> 
#> $`9`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd34a50>
#> 
#> $`10`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd34858>
#> 
#> $`11`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd34660>
#> 
#> $`12`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd34468>
#> 
#> $`13`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8be710>
#> 
#> $`14`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8be518>
#> 
#> $`15`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8be320>
#> 
#> $`16`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8be128>
#> 
#> $`17`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bdf30>
#> 
#> $`18`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bdd38>
#> 
#> $`19`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bdb40>
#> 
#> $`20`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bd948>
#> 
#> $`21`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bd750>
#> 
#> $`22`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bd558>
#> 
#> $`23`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bd360>
#> 
#> $`24`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bd168>
#> 
#> $`25`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bcf70>
#> 
#> $`26`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bcd78>
#> 
#> $`27`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bcb80>
#> 
#> $`28`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a8bc988>
#> 
#> $`29`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c141bd0>
#> 
#> $`30`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1419d8>
#> 
#> $`31`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1417e0>
#> 
#> $`32`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1415e8>
#> 
#> $`33`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1413f0>
#> 
#> $`34`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1411f8>
#> 
#> $`35`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c141000>
#> 
#> $`36`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c140e08>
#> 
#> $`37`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c140c10>
#> 
#> $`38`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c140a18>
#> 
#> $`39`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c140820>
#> 
#> $`40`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c140628>
#> 
#> $`41`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c140430>
#> 
#> $`42`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c140238>
#> 
#> $`43`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c140040>
#> 
#> $`44`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c13fe48>
#> 
#> $`45`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc88870>
#> 
#> $`46`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc88678>
#> 
#> $`47`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc88480>
#> 
#> $`48`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc88288>
#> 
#> $`49`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc88090>
#> 
#> $`50`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc87e98>
#> 
#> $`51`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc87ca0>
#> 
#> $`52`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc87aa8>
#> 
#> $`53`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc878b0>
#> 
#> $`54`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc87680>
#> 
#> $`55`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc87488>
#> 
#> $`56`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc87258>
#> 
#> $`57`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc87060>
#> 
#> $`58`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc86e68>
#> 
#> $`59`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bc86c70>
#> 
#> $`60`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba913e8>
#> 
#> $`61`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba911f0>
#> 
#> $`62`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90ff8>
#> 
#> $`63`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90e00>
#> 
#> $`64`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90c08>
#> 
#> $`65`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90a10>
#> 
#> $`66`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90818>
#> 
#> $`67`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90620>
#> 
#> $`68`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90428>
#> 
#> $`69`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90230>
#> 
#> $`70`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba90038>
#> 
#> $`71`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba8fe40>
#> 
#> $`72`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba8fc48>
#> 
#> $`73`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba8fa50>
#> 
#> $`74`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba8f858>
#> 
#> $`75`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6ba8f660>
#> 
#> $`76`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b3878>
#> 
#> $`77`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b3680>
#> 
#> $`78`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b3488>
#> 
#> $`79`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b3290>
#> 
#> $`80`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b3098>
#> 
#> $`81`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b2ea0>
#> 
#> $`82`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b2ca8>
#> 
#> $`83`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b2ab0>
#> 
#> $`84`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b28b8>
#> 
#> $`85`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b26c0>
#> 
#> $`86`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b24c8>
#> 
#> $`87`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b22d0>
#> 
#> $`88`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b20d8>
#> 
#> $`89`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b1ee0>
#> 
#> $`90`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b1ce8>
#> 
#> $`91`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c1b1af0>
#> 
#> $`92`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab58f8>
#> 
#> $`93`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab5700>
#> 
#> $`94`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab5508>
#> 
#> $`95`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab5310>
#> 
#> $`96`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab5118>
#> 
#> $`97`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab4f20>
#> 
#> $`98`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab4d28>
#> 
#> $`99`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab4b30>
#> 
#> $`100`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab4938>
#> 
#> $`101`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab4740>
#> 
#> $`102`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab4548>
#> 
#> $`103`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab4350>
#> 
#> $`104`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab4158>
#> 
#> $`105`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab3f60>
#> 
#> $`106`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bab3d68>
#> 
#> $`107`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51d0d0>
#> 
#> $`108`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51ced8>
#> 
#> $`109`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51cce0>
#> 
#> $`110`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51cab0>
#> 
#> $`111`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51c8b8>
#> 
#> $`112`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51c6c0>
#> 
#> $`113`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51c4c8>
#> 
#> $`114`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51c2d0>
#> 
#> $`115`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51c0a0>
#> 
#> $`116`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c51bea8>
#> 
#> $`117`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af6eef8>
#> 
#> $`118`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af6ed00>
#> 
#> $`119`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af6eb08>
#> 
#> $`120`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af6e910>
#> 
#> $`121`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af6e718>
#> 
#> $`2`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d7e00>
#> 
#> $`3`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d7bd0>
#> 
#> $`4`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d79d8>
#> 
#> $`5`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d77e0>
#> 
#> $`6`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d75e8>
#> 
#> $`7`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d73f0>
#> 
#> $`8`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d71f8>
#> 
#> $`9`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b5d7000>
#> 
#> $`10`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e7338>
#> 
#> $`11`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e7140>
#> 
#> $`12`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e6f48>
#> 
#> $`13`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e6d50>
#> 
#> $`14`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e6b58>
#> 
#> $`15`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e6960>
#> 
#> $`16`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e6768>
#> 
#> $`17`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e6570>
#> 
#> $`18`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e6378>
#> 
#> $`19`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e6180>
#> 
#> $`20`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e5f88>
#> 
#> $`21`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e5d90>
#> 
#> $`22`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e5b98>
#> 
#> $`23`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e59a0>
#> 
#> $`24`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b1e57a8>
#> 
#> $`25`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a619ea0>
#> 
#> $`26`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a619ca8>
#> 
#> $`27`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a619ab0>
#> 
#> $`28`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a619848>
#> 
#> $`29`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a619650>
#> 
#> $`30`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a619458>
#> 
#> $`31`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a619260>
#> 
#> $`32`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a619068>
#> 
#> $`33`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a618e70>
#> 
#> $`34`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a618c78>
#> 
#> $`35`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a618a80>
#> 
#> $`36`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a618888>
#> 
#> $`37`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a618690>
#> 
#> $`38`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a618498>
#> 
#> $`39`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a6182a0>
#> 
#> $`40`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a6180a8>
#> 
#> $`41`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70bc80>
#> 
#> $`42`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70ba50>
#> 
#> $`43`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70b858>
#> 
#> $`44`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70b628>
#> 
#> $`45`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70b430>
#> 
#> $`46`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70b238>
#> 
#> $`47`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70b040>
#> 
#> $`48`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70ae48>
#> 
#> $`49`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70ac50>
#> 
#> $`50`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70aa58>
#> 
#> $`51`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70a860>
#> 
#> $`52`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70a668>
#> 
#> $`53`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70a470>
#> 
#> $`54`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70a208>
#> 
#> $`55`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a70a010>
#> 
#> $`56`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d2f38>
#> 
#> $`57`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d2d40>
#> 
#> $`58`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d2b48>
#> 
#> $`59`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d2950>
#> 
#> $`60`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d2758>
#> 
#> $`61`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d2560>
#> 
#> $`62`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d2368>
#> 
#> $`63`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d2170>
#> 
#> $`64`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d1f78>
#> 
#> $`65`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d1d80>
#> 
#> $`66`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d1b88>
#> 
#> $`67`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d1990>
#> 
#> $`68`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d1798>
#> 
#> $`69`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d15a0>
#> 
#> $`70`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d13a8>
#> 
#> $`71`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c2d11b0>
#> 
#> $`72`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a4327c8>
#> 
#> $`73`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a4325d0>
#> 
#> $`74`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a4323d8>
#> 
#> $`75`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a4321e0>
#> 
#> $`76`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a431fe8>
#> 
#> $`77`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a431df0>
#> 
#> $`78`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a431bf8>
#> 
#> $`79`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a431a00>
#> 
#> $`80`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a431798>
#> 
#> $`81`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a4315a0>
#> 
#> $`82`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a4313a8>
#> 
#> $`83`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a4311b0>
#> 
#> $`84`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a430fb8>
#> 
#> $`85`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a430dc0>
#> 
#> $`86`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a430bc8>
#> 
#> $`87`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a4309d0>
#> 
#> $`88`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898ebd8>
#> 
#> $`89`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898e9e0>
#> 
#> $`90`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898e7e8>
#> 
#> $`91`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898e5f0>
#> 
#> $`92`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898e3f8>
#> 
#> $`93`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898e200>
#> 
#> $`94`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898e008>
#> 
#> $`95`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898de10>
#> 
#> $`96`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898dc18>
#> 
#> $`97`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898da20>
#> 
#> $`98`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898d828>
#> 
#> $`99`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898d630>
#> 
#> $`100`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898d438>
#> 
#> $`101`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898d240>
#> 
#> $`102`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6898d048>
#> 
#> $`103`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a508bb0>
#> 
#> $`104`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5089b8>
#> 
#> $`105`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5087c0>
#> 
#> $`106`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5085c8>
#> 
#> $`107`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a508360>
#> 
#> $`108`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a508168>
#> 
#> $`109`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a507f70>
#> 
#> $`110`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a507d78>
#> 
#> $`111`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a507b80>
#> 
#> $`112`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a507988>
#> 
#> $`113`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a507790>
#> 
#> $`114`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a507598>
#> 
#> $`115`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5073a0>
#> 
#> $`116`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a5071a8>
#> 
#> $`117`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a506fb0>
#> 
#> $`118`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6a506db8>
#> 
#> $`119`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b359ed0>
#> 
#> $`120`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b359cd8>
#> 
#> $`121`
#> function(x, y, N, M) cos((x-1)*pi*nm[1,1]/N)*sin((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b359ae0>
#> 
#> $`2`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b359258>
#> 
#> $`3`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b359060>
#> 
#> $`4`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b358e68>
#> 
#> $`5`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b358c70>
#> 
#> $`6`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b358a78>
#> 
#> $`7`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b358880>
#> 
#> $`8`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b358688>
#> 
#> $`9`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b358490>
#> 
#> $`10`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b358228>
#> 
#> $`11`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af74ef0>
#> 
#> $`12`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af74cf8>
#> 
#> $`13`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af74b00>
#> 
#> $`14`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af74908>
#> 
#> $`15`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af74710>
#> 
#> $`16`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af74518>
#> 
#> $`17`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af74320>
#> 
#> $`18`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af74128>
#> 
#> $`19`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af73f30>
#> 
#> $`20`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af73d38>
#> 
#> $`21`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af73b40>
#> 
#> $`22`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af73948>
#> 
#> $`23`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af73750>
#> 
#> $`24`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af73558>
#> 
#> $`25`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af73360>
#> 
#> $`26`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6af73168>
#> 
#> $`27`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48fa10>
#> 
#> $`28`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48f818>
#> 
#> $`29`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48f620>
#> 
#> $`30`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48f428>
#> 
#> $`31`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48f230>
#> 
#> $`32`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48f038>
#> 
#> $`33`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48ee40>
#> 
#> $`34`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48ec48>
#> 
#> $`35`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48ea50>
#> 
#> $`36`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48e858>
#> 
#> $`37`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48e5f0>
#> 
#> $`38`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48e3f8>
#> 
#> $`39`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48e200>
#> 
#> $`40`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48e008>
#> 
#> $`41`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48de10>
#> 
#> $`42`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b48dc18>
#> 
#> $`43`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18ebe0>
#> 
#> $`44`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18e9e8>
#> 
#> $`45`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18e7f0>
#> 
#> $`46`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18e5f8>
#> 
#> $`47`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18e400>
#> 
#> $`48`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18e208>
#> 
#> $`49`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18e010>
#> 
#> $`50`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18de18>
#> 
#> $`51`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18dc20>
#> 
#> $`52`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18da28>
#> 
#> $`53`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18d830>
#> 
#> $`54`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18d638>
#> 
#> $`55`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18d440>
#> 
#> $`56`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18d248>
#> 
#> $`57`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6b18d050>
#> 
#> $`58`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c396b88>
#> 
#> $`59`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c396990>
#> 
#> $`60`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c396798>
#> 
#> $`61`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c3965a0>
#> 
#> $`62`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c3963a8>
#> 
#> $`63`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c3961b0>
#> 
#> $`64`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c395fb8>
#> 
#> $`65`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c395d50>
#> 
#> $`66`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c395b58>
#> 
#> $`67`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c395960>
#> 
#> $`68`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c395768>
#> 
#> $`69`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c395570>
#> 
#> $`70`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c395378>
#> 
#> $`71`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c395180>
#> 
#> $`72`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c394f88>
#> 
#> $`73`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c394d90>
#> 
#> $`74`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e13758>
#> 
#> $`75`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e13560>
#> 
#> $`76`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e13368>
#> 
#> $`77`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e13170>
#> 
#> $`78`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e12f78>
#> 
#> $`79`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e12d80>
#> 
#> $`80`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e12b88>
#> 
#> $`81`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e12990>
#> 
#> $`82`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e12798>
#> 
#> $`83`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e125a0>
#> 
#> $`84`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e123a8>
#> 
#> $`85`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e121b0>
#> 
#> $`86`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e11fb8>
#> 
#> $`87`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e11dc0>
#> 
#> $`88`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e11bc8>
#> 
#> $`89`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac67e119d0>
#> 
#> $`90`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c477628>
#> 
#> $`91`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c477430>
#> 
#> $`92`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c4771c8>
#> 
#> $`93`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c476fd0>
#> 
#> $`94`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c476dd8>
#> 
#> $`95`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c476be0>
#> 
#> $`96`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c4769b0>
#> 
#> $`97`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c4767b8>
#> 
#> $`98`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c476588>
#> 
#> $`99`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c476390>
#> 
#> $`100`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c476198>
#> 
#> $`101`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c475fa0>
#> 
#> $`102`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c475da8>
#> 
#> $`103`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c475bb0>
#> 
#> $`104`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6c4759b8>
#> 
#> $`105`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4ee90>
#> 
#> $`106`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4ec98>
#> 
#> $`107`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4eaa0>
#> 
#> $`108`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4e8a8>
#> 
#> $`109`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4e6b0>
#> 
#> $`110`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4e4b8>
#> 
#> $`111`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4e2c0>
#> 
#> $`112`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4e0c8>
#> 
#> $`113`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4ded0>
#> 
#> $`114`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4dcd8>
#> 
#> $`115`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4dae0>
#> 
#> $`116`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4d8e8>
#> 
#> $`117`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4d6f0>
#> 
#> $`118`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4d4f8>
#> 
#> $`119`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4d300>
#> 
#> $`120`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6be4d108>
#> 
#> $`121`
#> function(x, y, N, M) sin((x-1)*pi*nm[1,1]/N)*cos((y-1)*pi*nm[1,2]/M)
#> <environment: 0x55ac6bd36110>
#> 

polynomial_2d(c(3,3), dim(Z_potts))

#> $`2`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb524a0>
#> 
#> $`3`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb522a8>
#> 
#> $`4`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb520b0>
#> 
#> $`5`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb51eb8>
#> 
#> $`6`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb51cc0>
#> 
#> $`7`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb558f8>
#> 
#> $`8`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb55700>
#> 
#> $`9`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb55508>
#> 
#> $`10`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb55310>
#> 
#> $`11`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb55118>
#> 
#> $`12`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb54f20>
#> 
#> $`13`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb54d28>
#> 
#> $`14`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb54b30>
#> 
#> $`15`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb54938>
#> 
#> $`16`
#> function(x, y, N, M) return((x - N/2)^nm[1,1]*(y - M/2)^nm[1,2])
#> <environment: 0x55ac6bb54740>
#>