BOLD5000 neuroimaging data

An image extracted from the "BOLD5000" open dataset. It was read from the file in path BOLD5000/DS001499/SUB-CSI2/SES-16/ANAT/SUB-CSI2_SES-16_T1W.NII.GZ, available at the OpenNeuro platform (https://openneuro.org/datasets/ds001499/versions/1.3.0).

bold5000

Format

An object of class matrix (inherits from array) with 176 rows and 256 columns.

Details

The file was read using the oro.nifti package and the image was extracted from the matrix in slice 160.

References

Chang, N., Pyles, J. A., Marcus, A., Gupta, A., Tarr, M. J., & Aminoff, E. M. (2019). BOLD5000, a public fMRI dataset while viewing 5000 visual images. Scientific data, 6(1), 1-18.