Stochastic Approximation algorithm for MRFs on 2d lattices

Estimates the parameters of a MRF by successively sampling from the a parameter and updating it by comparing the sufficient statistics of the sampled field and the reference field.

fit_sa(Z, mrfi, family = "onepar", gamma_seq, init = 0, cycles = 5,
  refresh_each = length(gamma_seq) + 1, refresh_cycles = 60)

Arguments

Z
A matrix object containing the observed MRF.

mrfi
A mrfi object representing the interaction structure.

family
The family of parameter restrictions to potentials. Families are: 'onepar', 'oneeach', 'absdif', 'dif' or 'free'. See mrf2d-familiy.

gamma_seq
A numeric vector with the sequence of constants used in each step \(\gamma_t\).

init
The initial value to be used in the optimization. It can be:

A valid array of parameter values according to family.

0. If set to 0 an array with `0`` in all entries is created.

refresh_each
An integer with the number of iterations taken before a complete refresh (restart from a random state). This prevents the sample from being stuck in a mode for too long. Defaults to length(gamma_seq) + 1 (no refresh happens).

refresh_cycles
An iteger indicating how many Gibbs Sampler cycles are performed when a refresh happens. Larger is usually better, but slower.

Value

A list object with the following elements:

theta: The estimated array of potentials.

metrics: A data.frame containing the the euclidean distance between the sufficient statics computed for Z and the current sample.

Details

Examples where Stochastic Approximation is used with MRFs are (Gimel'farb 1996) , (Atchadé et al. 2013) .

$$\theta_{t+1} = \theta_t - \gamma_t(T(Z_t) - T(Z)),$$ where \(T(Z)\) is the sufficient statistics for the reference field, \(T(Z_t)\) is the sufficient statistics for a field sampled from \(\theta_t\).

Note

Stochastic Approximation is called "Controllable Simulated Annealing" in some references.

References

Wikipedia (2019). “Stochastic approximation.” https://en.wikipedia.org/wiki/Stochastic_approximation.

Atchadé YF, Lartillot N, Robert C, others (2013). “Bayesian computation for statistical models with intractable normalizing constants.” Brazilian Journal of Probability and Statistics, 27(4), 416--436.

Gimel'farb GL (1996). “Texture modeling by multiple pairwise pixel interactions.” IEEE Transactions on pattern analysis and machine intelligence, 18(11), 1110--1114.

Examples

# NOT RUN {
set.seed(2)
# }