Pseudo-likelihood function for MRFs on 2d lattices

Computes the pseudo-likelihood function of a Markov Random Field on a 2-dimensional lattice.

pl_mrf2d(Z, mrfi, theta, log_scale = TRUE)

Arguments

Z
A matrix with integers in {0,...,C}.

mrfi
A mrfi object representing the interaction structure.

theta
A 3-dimensional array describing potentials. Slices represent interacting positions, rows represent pixel values and columns represent neighbor values. As an example: theta[1,3,2] has the potential pairs of values 0,2 in the second relative position of mrfi.

log_scale
A logical value indicatin g whether the returned value should be in logarithmic scale.

Value

A numeric with the pseudo-likelihood value.

Details

The pseudo-likelihood function is defined as the product of conditional distributions:

$$\prod_i P(Z_i | Z_{{N}_i}, \theta).$$

For more details see the guide vignette: vignette("guide", package = "mrf2d")

Examples

pl_mrf2d(Z_potts, mrfi(1), theta_potts)

#> [1] -10708.08