EM estimation for Gaussian Hidden Markov field

fit_ghm fits a Gaussian Mixture model with hidden components driven by a Markov random field with known parameters. The inclusion of a linear combination of basis functions as a fixed effect is also possible.

The algorithm is an implementation of (Zhang et al. 2001) .

fit_ghm(Y, mrfi, theta, fixed_fn = list(), equal_vars = TRUE,
  init_mus = NULL, init_sigmas = NULL, maxiter = 100,
  max_dist = 10^-3, icm_cycles = 6, verbose = TRUE)

Arguments

Y
A matrix of observed (continuous) pixel values.

mrfi
A mrfi object representing the interaction structure.

theta
A 3-dimensional array describing potentials. Slices represent interacting positions, rows represent pixel values and columns represent neighbor values. As an example: theta[1,3,2] has the potential pairs of values 0,2 in the second relative position of mrfi.

fixed_fn
A list of functions fn(x,y) to be considered as a fixed effect. See basis_functions.

equal_vars
logical indicating if the mixture model has the same variances in all mixture components.

init_mus
Optional. A numeric with length (C+1) with the initial mean estimate for each component.

init_sigmas
Otional. A numeric with length (C+1) with initial sample deviation estimate for each component.

maxiter
The maximum number of iterations allowed. Defaults to 100.

max_dist
Defines a stopping condition. The algorithm stops if the maximum absolute difference between parameters of two consecutive iterations is less than max_dist.

icm_cycles
Number of steps used in the Iterated Conditional Modes algorithm executed in each interaction. Defaults to 6.

verbose
logical indicating wheter to print the progress or not.

Value

A list containing:

par: A data.frame with \(\mu\) and \(\sigma\) estimates for each component.

fixed: A matrix with the estimated fixed effect in each pixel.

Z_pred: A matrix with the predicted component (highest probability) in each pixel.

predicted: A matrix with the fixed effect + the \(\mu\) value for the predicted component in each pixel.

iterations: Number of EM iterations done.

Details

If either init_mus or init_sigmas is NULL an EM algorithm considering an independent uniform distriburion for the hidden component is fitted first and its estimated means and sample deviations are used as initial values. This is necessary because the algorithm may not converge if the initial parameter configuration is too far from the maximum likelihood estimators.

max_dist defines a stopping condition. The algorithm will stop if the maximum absolute difference between (\(\mu\) and \(\sigma\)) parameters in consecutive iterations is less than max_dist.

$$max{max | \mu_{t+1} - \mu_t |, | \sigma{t+1} - \sigma_t|} <= max_dist$$

Note that if the maximum probability configuration for Z is unchanged between two iterations, then the parameter estimates will also be unchanged, this makes possible to set max_dist = 0 and hit the stopping condition.

For more details see the guide vignette: vignette("guide", package = "mrf2d")

References

Zhang Y, Brady M, Smith S (2001). “Segmentation of brain MR images through a hidden Markov random field model and the expectation-maximization algorithm.” IEEE transactions on medical imaging, 20(1), 45--57.

Examples

# Sample a Gaussian mixture with components given by Z_potts
# mean values are 0, 1 and 2 and a linear effect on the x-axis.

# NOT RUN {
set.seed(2)
Y <- Z_potts + rnorm(length(Z_potts), sd = 0.4) + row(Z_potts)*0.01
# Check what the data looks like
cplot(Y)

fixed <- polynomial_2d(c(1,1), dim(Y))
fit <- fit_ghm(Y, mrfi = mrfi(1), theta = theta_potts, fixed_fn = fixed)
fit$par
cplot(fit$fixed)
# }