Maximum Pseudo-likelihood fitting of MRFs on 2d lattices.

Parameter estimation for Markov random fields via Pseudo-Likelihood function optimization. See pl_mrf2d for information on the Pseudo-Likelihood function.

fit_pl(Z, mrfi, family = "onepar", init = 0, optim_args = list(method
  = "BFGS"), return_optim = FALSE)

Arguments

Z
A matrix object containing the observed MRF.

mrfi
A mrfi object representing the interaction structure.

family
The family of parameter restrictions to potentials. Families are: 'onepar', 'oneeach', 'absdif', 'dif' or 'free'. See mrf2d-familiy.

init
The initial value to be used in the optimization. It can be:

A valid array of parameter values according to family.

0. If set to 0 an array with `0`` in all entries is created.

optim_args
Additional parameters passed to optim().

return_optim
logical indicating whether information from the optim() call are returned.

Value

A list object with elements:

theta: The estimated array of potential values.

value: The optimal pseudo-likelihood value.

opt.xxx(if return_optim is TRUE): Information returned by the optim() function used for the optimization.

Examples

fit_pl(Z_potts, mrfi(1), family = "onepar")

#> $theta
#> , , 1
#> 
#>            0          1          2
#> 0  0.0000000 -0.9906675 -0.9906675
#> 1 -0.9906675  0.0000000 -0.9906675
#> 2 -0.9906675 -0.9906675  0.0000000
#> 
#> , , 2
#> 
#>            0          1          2
#> 0  0.0000000 -0.9906675 -0.9906675
#> 1 -0.9906675  0.0000000 -0.9906675
#> 2 -0.9906675 -0.9906675  0.0000000
#> 
#> 
#> $value
#> [1] -10707.56
#> 

fit_pl(Z_potts, mrfi(1), family = "oneeach")

#> $theta
#> , , 1
#> 
#>            0          1          2
#> 0  0.0000000 -0.9635956 -0.9635956
#> 1 -0.9635956  0.0000000 -0.9635956
#> 2 -0.9635956 -0.9635956  0.0000000
#> 
#> , , 2
#> 
#>           0         1         2
#> 0  0.000000 -1.017917 -1.017917
#> 1 -1.017917  0.000000 -1.017917
#> 2 -1.017917 -1.017917  0.000000
#> 
#> 
#> $value
#> [1] -10705.05
#> 

fit_pl(Z_potts, mrfi(2), family = "onepar")

#> $theta
#> , , 1
#> 
#>            0          1          2
#> 0  0.0000000 -0.3355747 -0.3355747
#> 1 -0.3355747  0.0000000 -0.3355747
#> 2 -0.3355747 -0.3355747  0.0000000
#> 
#> , , 2
#> 
#>            0          1          2
#> 0  0.0000000 -0.3355747 -0.3355747
#> 1 -0.3355747  0.0000000 -0.3355747
#> 2 -0.3355747 -0.3355747  0.0000000
#> 
#> , , 3
#> 
#>            0          1          2
#> 0  0.0000000 -0.3355747 -0.3355747
#> 1 -0.3355747  0.0000000 -0.3355747
#> 2 -0.3355747 -0.3355747  0.0000000
#> 
#> , , 4
#> 
#>            0          1          2
#> 0  0.0000000 -0.3355747 -0.3355747
#> 1 -0.3355747  0.0000000 -0.3355747
#> 2 -0.3355747 -0.3355747  0.0000000
#> 
#> , , 5
#> 
#>            0          1          2
#> 0  0.0000000 -0.3355747 -0.3355747
#> 1 -0.3355747  0.0000000 -0.3355747
#> 2 -0.3355747 -0.3355747  0.0000000
#> 
#> , , 6
#> 
#>            0          1          2
#> 0  0.0000000 -0.3355747 -0.3355747
#> 1 -0.3355747  0.0000000 -0.3355747
#> 2 -0.3355747 -0.3355747  0.0000000
#> 
#> 
#> $value
#> [1] -12072.36
#>