Parameter restriction families

Different parameter restrictions can be included in estimation processes to make sure mrf2d can sucessfully include a wide range of models available in its inference functions. For a more complete description of which restrictions are used to characterize some of the most commom models read the corresponding section of the package's vignette

For model identifiability, at least one linear restriction is necessary. mrf2d always assume \(\theta_{0,0,r} = 0\) for all relative positions \(r\).

Additionally, each family of restrictions may introduce other restrictions:

'onepar'

This family assumes the model is defined by a single parameter by adding the restriction

$$\theta_{a,b,r} = \gamma * 1(a != b).$$

Here \(1()\) denotes de indicator function. In words, the parameter must be the same value for any pair with different values and 0 for any equal-valued pair.

'oneeach'

Similar to 'onepar', parameters are 0 for equal-valued pairs and a constant for pairs with different values, but the constant may differ between different relative positions \(r\):

$$\theta{a,b,r} = \gamma_r * 1(a != b).$$

'absdif'

All parameters \(\theta_{a,b,r}\) with the same absolute difference between \(a\) and \(b\) must be equal within each relative position \(r\).

$$\theta_{a,b,r} = \sum_d \gamma_{d,r} * 1(|a-b| == d)$$

'dif'

The same as 'absdif', but parameters may differ between positive and negative differences.

$$\theta_{a,b,r} = \sum_d \gamma_{d,r} * 1(a-b == d)$$

'free'

No additional restriction, all parameters other than \(\theta_{0,0,r}\) vary freely.